refactor: UserRepository를 인터페이스로 분리
- UserRepository는 String id로 User를 찾아오며, ArticleRepository는 Long id로
Article을 찾아오기 때문에 분리를 시도함
